requirements.txt
python-telegram-bot>=21.0
SQLAlchemy>=2.0
aiosqlite>=0.19
